server/requirements.txt
Flask==2.1.0
Flask-SocketIO==5.1.1
Flask-CORS==3.0.10
python-dotenv==0.19.2
gunicorn==20.1.0
eventlet==0.33.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3
erlang==3.5.0